Windows port, changed long long to int64_t in logfilemanager
git-svn-id: file:///svn/toku/tokudb@14387 c7de825b-a66e-492c-adef-691d508d4ae1

+++ newbrt/logfilemgr.h
@@ -9,7 +9,7 @@
 
 // this is the basic information we need to keep per logfile
 struct toku_logfile_info {
-    long long index;
+    int64_t index;
     LSN maxlsn;
 };
 typedef struct toku_logfile_info *TOKULOGFILEINFO;
